Bright-Coffee-Shop-Analysis
Added SQL queries for exploratory data analysis including checks on store locations, product categories, transaction dates, and sales metrics.

The coffee shop industry in one of the most difficult businesses to turn a consistent profit in. Costs are high, margins are narrow, and competition is constant. It appears that the coffee shops are doing pretty well towards the winter seasons which makes sense as the weather is usually cold and that is the time were people appreciate warm bevarages.

To keep the revenue afloat across all seasons, we can develop a high-margin breakfast offering.
Breakfast is one of the most cost-effective categories in food service. Items like muffins, croissants, egg sandwiches, and bagels are inexpensive to produce and pair well with coffee. They require relatively simple preparation and can be sold at a strong mark.